+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 
 ### Other
 - Added JetBrains Marketplace error reporting [#1039](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1027) for Issue [1026](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1039)
+- Retrieve SOLR Cores view background process [#1082](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1082)
 
 ## [2024.1.0]